[noref] Refactor for bootstrap only in Z7
