--- client/chrome.manifest
@@ -1,3 +1,0 @@
-content   scite-zotero-plugin             chrome/content/
-locale scite-zotero-plugin en-US     chrome/locale/en-US/
-skin      scite-zotero-plugin   default   chrome/skin/
