--- client/locale/en-US/scite-zotero-plugin.ftl
@@ -1,2 +1,0 @@
-make-it-green-instead =
-    .label = Make It Green Instead
