--- client/skin/html_overlay.css
@@ -1,42 +1,0 @@
-button *, button:focus *
-{
-  background: transparent !important;
-  background-image: none !important;
-  border: 0 !important;
-  box-shadow: none !important;
-}
-
-.scite-cell {
-  display: flex;
-  align-items: center;
-}
-
-.scite-cell .icon.icon-bg.cell-icon{
-  margin-right: 6px;
-  width: 14px;
-  height: 15px;
-}
-.zotero-items-column-supporting .icon.icon-bg.cell-icon
-{
-  background-image: url('supporting.png');
-}
-
-.zotero-items-column-contrasting .icon.icon-bg.cell-icon
-{
-  background-image: url('contrasting.png');
-}
-
-.zotero-items-column-mentioning .icon.icon-bg.cell-icon
-{
-  background-image: url('mentioning.png');
-}
-
-.zotero-items-column-loading.icon.icon-bg.cell-icon
-{
-  background-image: url('loading.jpg');
-}
-
-.zotero-items-column-citingPublications .icon.icon-bg.cell-icon
-{
-  background-image: url('total_publications.png');
-}
