--- client/skin/xul_overlay.css
@@ -1,50 +1,0 @@
-@namespace url("http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul");
-
-button *, button:focus *
-{
-  background: transparent !important;
-  background-image: none !important;
-  border: 0 !important;
-  box-shadow: none !important;
-}
-
-#zotero-items-tree treechildren::-moz-tree-image(scite-state-supporting)
-{
-  margin: 1px 5px 0;
-  list-style-image: url('supporting.png');
-  height: 10px;
-}
-
-#zotero-items-tree treechildren::-moz-tree-image(scite-state-contrasting)
-{
-  margin: 1px 5px 0;
-  list-style-image: url('contrasting.png');
-  height: 10px;
-}
-
-#zotero-items-tree treechildren::-moz-tree-image(scite-state-mentioning)
-{
-  margin: 1px 5px 0;
-  list-style-image: url('mentioning.png');
-  height: 10px;
-}
-
-#zotero-items-tree treechildren::-moz-tree-image(scite-state-loading)
-{
-  margin: 1px 5px 0;
-  list-style-image: url('loading.jpg');
-  height: 10px;
-}
-
-#zotero-items-tree treechildren::-moz-tree-image(scite-state-total)
-{
-  margin: 1px 5px 0;
-  height: 10px;
-}
-
-#zotero-items-tree treechildren::-moz-tree-image(scite-state-citingPublications)
-{
-  margin: 1px 5px 0;
-  list-style-image: url('total_publications.png');
-  height: 10px;
-}
