+++ tsconfig.json
@@ -19,7 +19,7 @@
       "./node_modules/@types"
     ]
   },
-  "include": [ "*.ts", "chrome/content", "client/chrome/content/**/*.ts", "typing", "node_modules/zotero-types", "client/debug.ts" ],
+  "include": [ "*.ts", "chrome/content", "client/chrome/content/**/*.ts", "typing", "node_modules/zotero-types", "client/content/debug.ts" ],
   "exclude": [
     "node_modules",
     "**/*.spec.ts",
